% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM8Kspec-package.R
\docType{package}
\name{AM8Kspec-package}
\alias{AM8Kspec}
\alias{AM8Kspec-package}
\title{AM8Kspec: Making life easier with our spectrophotometer}
\description{
Read and parse the text output from our spectrophotometer, recreate calibrations, diagnose measurements and quality.
}
\author{
\strong{Maintainer}: Zach Simpson \email{zachary.simpson@usda.gov}

}
\keyword{internal}
