% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spec_calib.R
\name{extract_spec_calib}
\alias{extract_spec_calib}
\title{Extract the calibration data from raw spec file.}
\usage{
extract_spec_calib(spec_raw)
}
\arguments{
\item{spec_raw}{The output from \code{read_raw_spec_txt}}
}
\value{
A tibble with the standards' IDs, absorbances, and concentrations.
}
\description{
Extract the calibration data from raw spec file.
}
