% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_spec_txt.R
\name{read_raw_spec_txt}
\alias{read_raw_spec_txt}
\title{Utility function to read in the raw text output from the machine & standardise it. This is just an initial step before prettying it up.}
\usage{
read_raw_spec_txt(filename)
}
\arguments{
\item{filename}{The txt file to read from, given in the machine's lovely format.}
}
\value{
A standardised tibble with just 3 columns -- these are for subsequent use.
}
\description{
Utility function to read in the raw text output from the machine & standardise it. This is just an initial step before prettying it up.
}
